% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/righthojeeves.R
\docType{data}
\name{adamselindistress}
\alias{adamselindistress}
\title{"A Damsel in Distress"}
\format{A character vector with 10291 elements}
\source{
\url{https://www.gutenberg.org/files/2233/2233-8.txt}
}
\usage{
adamselindistress
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
novel "A Damsel in Distress".
}
\keyword{datasets}
