% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{amanofmeans}
\alias{amanofmeans}
\title{A Man of Means}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/files/8713/8713.txt}
}
\usage{
amanofmeans
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"A Man of Means", a collection of six short stories published in 1914,
starring Roland Beake, whose financial success is not always a blessing.
}
\keyword{datasets}
