% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{aprefectsuncle}
\alias{aprefectsuncle}
\title{A Prefect's Uncle}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/files/6985/6985.txt}
}
\usage{
aprefectsuncle
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"A Prefect's Uncle", a collection of stories set in a private school for boys.
}
\keyword{datasets}
