% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{awodehousemiscellany}
\alias{awodehousemiscellany}
\title{"A Wodehouse Miscellany"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/8190/pg8190.txt}
}
\usage{
awodehousemiscellany
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"A Wodehouse Miscellany", a collection of articles, poems, and stories.
}
\keyword{datasets}
