% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{copyAlteryxRPlugin}
\alias{copyAlteryxRPlugin}
\title{Copy predictive macros, samples and plugins from SVN}
\usage{
copyAlteryxRPlugin(to = NULL, svnDir = getOption("alteryx.svndir"))
}
\arguments{
\item{to}{directory to copy files to.}

\item{svnDir}{path to svn branch.}
}
\description{
Copy predictive macros, samples and plugins from SVN
}

