% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDashboard.R
\name{createTestDashboard}
\alias{createTestDashboard}
\title{Create a dashboard of test results}
\usage{
createTestDashboard(testDirs)
}
\arguments{
\item{testDirs}{directories containing the tests.}
}
\description{
Create a dashboard of test results
}

