% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{deathattheexcelsiorandotherstories}
\alias{deathattheexcelsiorandotherstories}
\title{Death at The Excelsior and Other Stories}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\usage{
deathattheexcelsiorandotherstories
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"Death at the Excelsior and Other Stories", a postumously published collection of short
stories.

#' @source \url{http://www.gutenberg.org/cache/epub/8176/pg8176.txt}
}
\keyword{datasets}
