% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractAnnotationConfig}
\alias{extractAnnotationConfig}
\title{Extract annotation from macro configuration}
\usage{
extractAnnotationConfig(pluginName)
}
\arguments{
\item{pluginName}{path to yxmc file}
}
\description{
Extract annotation from macro configuration
}
\seealso{
Other extract: \code{\link{extractAnnotationInputOutput}},
  \code{\link{extractMacroConstants}},
  \code{\link{extractQuestionConstants}}
}
