% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractAnnotationInputOutput}
\alias{extractAnnotationInputOutput}
\alias{extractAnnotationInput}
\alias{extractAnnotationOutput}
\title{Extract annotation from macro inputs and outputs}
\usage{
extractAnnotationInputOutput(pluginName, type = "input")

extractAnnotationInput(pluginName)

extractAnnotationOutput(pluginName)
}
\arguments{
\item{pluginName}{path to yxmc file}

\item{type}{whether to extract from input or output}
}
\description{
Extract annotation from macro inputs and outputs
}
\seealso{
Other extract: \code{\link{extractAnnotationConfig}},
  \code{\link{extractMacroConstants}},
  \code{\link{extractQuestionConstants}}

Other extract: \code{\link{extractAnnotationConfig}},
  \code{\link{extractMacroConstants}},
  \code{\link{extractQuestionConstants}}

Other extract: \code{\link{extractAnnotationConfig}},
  \code{\link{extractMacroConstants}},
  \code{\link{extractQuestionConstants}}
}
