% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractMacroConstants}
\alias{extractMacroConstants}
\title{Extract macro constants.}
\usage{
extractMacroConstants(root, yxmc = NULL)
}
\arguments{
\item{root}{xml object containing the root node of a yxmc file.}

\item{yxmc}{path to yxmc file.}
}
\description{
These values if present should override the defaults set in the macro.
}
\seealso{
Other extract: \code{\link{extractAnnotationConfig}},
  \code{\link{extractAnnotationInputOutput}},
  \code{\link{extractQuestionConstants}}
}
