% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractQuestionConstants}
\alias{extractQuestionConstants}
\title{Extract question constants and return code that can be inserted}
\usage{
extractQuestionConstants(macro, input = "config")
}
\arguments{
\item{macro}{path to macro from which to extract question constants}

\item{input}{variable to assign the question constants to}
}
\description{
Extract question constants and return code that can be inserted
}
\seealso{
Other extract: \code{\link{extractAnnotationConfig}},
  \code{\link{extractAnnotationInputOutput}},
  \code{\link{extractMacroConstants}}
}
