% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnfunctions.R
\name{getAyxSvnRDirs}
\alias{getAyxSvnRDirs}
\title{Get Alteryx R SVN Directories}
\usage{
getAyxSvnRDirs(svnDir = getOption("alteryx.svndir"), rVersion = NULL)
}
\arguments{
\item{svnDir}{svn directory}

\item{rVersion}{optional argument to specify an R version.}
}
\value{
a named list containing paths to the R library and
  \code{3rdParty//R//Installer} on SVN
}
\description{
Get Alteryx R SVN Directories
}

