% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFromGithub.R
\name{getNewDependencies}
\alias{getNewDependencies}
\title{Get new dependencies for a package}
\usage{
getNewDependencies(pkg)
}
\arguments{
\item{pkg}{package to get new dependencies for}
}
\description{
Get new dependencies for a package
}

