% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HammingMatrix.R
\name{HammingMatrix}
\alias{HammingMatrix}
\title{HammingMatrix}
\usage{
HammingMatrix(aln)
}
\arguments{
\item{aln}{Nucleotide sequence alignment of class DNAbin or phyDat}

\item{snp_threshold}{Integer (n>=0) indicating the maximum within-group
Hamming distance.}
}
\value{
A tibble with columns sequence_name and sequence_group.
}
\description{
Calculate Hamming distance of a nucleotide sequence alignment. This is a
dependency for GroupSequences.
}
