% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{indiscretionsofarchie}
\alias{indiscretionsofarchie}
\title{Indiscretions of Archie}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\usage{
indiscretionsofarchie
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"Indiscretions of Archie", a hilarious narrative documenting Archie's travails in
America, chasing love.

#' @source \url{http://www.gutenberg.org/files/3756/3756-0.txt}
}
\keyword{datasets}
