% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{install_all_pkgs}
\alias{install_all_pkgs}
\title{Install All Packages Needed for Alteryx Predictive Tools}
\usage{
install_all_pkgs(currentRVersion, installation = c("dev", "svn"),
  readmeManifest = TRUE, dataXPath = NULL,
  repos = "https://cloud.r-project.org", useGitHub = FALSE,
  ayxDepend = "stringr")
}
\arguments{
\item{installation}{One of "dev" or "svn". In the case of "dev", the
needed CRAN packages are installed into the system library of the user's
development installation of R. When "svn" is selected, then the packages
are installed to the system library of the R installation located in the
user's local copy of the relevant SVN repository. The development R
The respository's path is determined by the alteryx.svndir global options
setting.}

\item{readmeManifest}{A logical flag indicating whether the Readme file
and the manifest file are saved after installing all the
needed package. This is only relevant for installing packages into the SVN
R installation.}

\item{dataXPath}{The local full path to an appropriate binary installer of
the AlteryxRDataX package. If its value is NULL, then no attempt will be
made to install the package.}

\item{repos}{The CRAN repository to use for package installation. The
default is https://cloud.r-project.org.}

\item{useGitHub}{Install the Alteryx predictive packages other than
AlteryxRDataX from Alteryx's CRAN like repository on GitHub.
The default is FALSE.}

\item{ayxDepend}{A character vector of CRAN packages that Alteryx packages
depend on, but are not a dependency of other CRAN packages. Currently this
is just the package stringr.}
}
\description{
The function can be used to install Altery's R packages to either a
user's development R installation or the R installation in the user's
local copy of the SVN repository of an Alteryx development branch.
NOTE: To use this function, the R session being used must be running
in administrator mode to allow for appropriate read/write permissions.
}
