% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{install_CRAN_pkgs}
\alias{install_CRAN_pkgs}
\title{Install CRAN Packages needed for Alteryx predictive tools}
\usage{
install_CRAN_pkgs(currentRVersion, installation = c("dev", "svn"),
  repos = "https://cloud.r-project.org")
}
\arguments{
\item{currentRVersion}{The current version of R being used by Alteryx's
predictive tools.}

\item{installation}{One of "dev" or "svn". In the case of "dev", the
needed CRAN packages are installed into the system library of the user's
development installation of R. When "svn" is selected, then the packages
are installed to the system library of the R installation located in the
user's local copy of the relevant SVN repository. The development R
The respository's path is determined by the alteryx.svndir global options
setting.}

\item{repos}{The CRAN repository to use for package installation. The
default is https://cloud.r-project.org.}
}
\description{
The function can be used to install needed CRAN packages for the predictive
tools to either a user's development R installation or the R installation in
the user's local copy of the SVN repository of an Alteryx development
branch. NOTE: To use this function, the R session being used must be running
in administrator mode to allow for appropriate read/write permissions.
}
