% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{installAllPackages}
\alias{installAllPackages}
\title{Install all needed packages that are missing}
\usage{
installAllPackages(dev = TRUE)
}
\arguments{
\item{dev}{boolean indicating if dev versions of packages should be installed.}
}
\description{
Install all needed packages that are missing
}

