% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{installAllPackages2}
\alias{installAllPackages2}
\title{Install All Needed Packages V2}
\usage{
installAllPackages2(branch = "Predictive_Dev", buildDir = NULL,
  ayxRepo = "https://alteryx.github.io/drat",
  buildRepo = "\\\\\\\\DEN-IT-FILE-07\\\\BuildRepo")
}
\arguments{
\item{branch}{string indicating svn branch.}

\item{buildDir}{build directory.}

\item{ayxRepo}{string indicating cran-like repo for alteryx packages}

\item{buildRepo}{build repo.}
}
\description{
Alteryx packages, with the exception of AlteryxRDataX, are installed from
the Alteryx drat repo on GitHub, while AlteryxRDataX is installed from
either the binary installer of the package of the most recent nightly
build of the specified branch, or from a local a local directory. The local
directory option would allow for an installation of AlteryxRDataX from
source.
}
