% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{installAlteryx}
\alias{installAlteryx}
\title{Install Alteryx}
\usage{
installAlteryx(installers)
}
\arguments{
\item{installers}{list of paths to named installers.}
}
\description{
Install Alteryx and optionally, the Predictive Tools using the
command line SDK. It does a silent install skipping through all dialogs and
accepting all defaults.
}

