% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnfunctions.R
\name{installToSvn}
\alias{installToSvn}
\title{Install packages to R library in SVN}
\usage{
installToSvn(srcPkg, to_install = TRUE, rVersion = getRversion())
}
\arguments{
\item{srcPkg}{path to source package}

\item{to_install}{whether or not to actually install dependencies}

\item{rVersion}{version of R}
}
\description{
Install packages to R library in SVN
}

