% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{package}
\name{jeeves}
\alias{jeeves}
\alias{wodehouse}
\alias{pgwodehouse}
\alias{psmith}
\alias{jeeves-package}
\title{Selected novels of P. G. Wodehouse}
\description{
This package contains the complete text of selected novels of
P. G. Wodehouse formatted to be convenient for text analysis.
}
