% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{jillthereckless}
\alias{jillthereckless}
\title{"Jill the Reckless"}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14973 rows and 3 columns.}
\source{
\url{http://www.gutenberg.org/cache/epub/20533/pg20533.txt}
#' @format A `tbl_df`, `tbl`, and `data.frame`
}
\usage{
jillthereckless
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
1920 classic love story that spans the Atlantic.
}
\keyword{datasets}
