% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-layout.R
\name{jvAccordionItem}
\alias{jvAccordionItem}
\title{Create an accordion item}
\usage{
jvAccordionItem(title, ..., id = makeHtmlId(title))
}
\arguments{
\item{title}{the title for the accordion item}

\item{...}{elements to build the accordion item}

\item{id}{html id for the accordion item}
}
\description{
Create an accordion item
}
