% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-layout.R
\name{jvContent}
\alias{jvContent}
\title{Create a content div}
\usage{
jvContent(..., .list = NULL)
}
\arguments{
\item{...}{elements to pass.}

\item{.list}{same as ..., but passed as a list.}
}
\description{
Create a content div
}
