% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-layout.R
\name{jvCustomizePage}
\alias{jvCustomizePage}
\title{Create the customize page}
\usage{
jvCustomizePage(title, ...)
}
\arguments{
\item{title}{tile of the page}

\item{...}{elements to build the page from}
}
\description{
Create the customize page
}
