% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-layout.R
\name{jvMakeDataItemsToInitialize}
\alias{jvMakeDataItemsToInitialize}
\title{Create data items to initialize}
\usage{
jvMakeDataItemsToInitialize(pluginDir = ".", curPage, curTab = NULL,
  curToggle = NULL, ...)
}
\arguments{
\item{pluginDir}{path to plugin directory}

\item{curPage}{id of page to be active by default}

\item{curTab}{id of tab to be active by default}

\item{curToggle}{id of toggle to be active by default}

\item{...}{other data items to be initialized before load}
}
\description{
Create data items to initialize
}
