% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-layout.R
\name{jvSetupPage}
\alias{jvSetupPage}
\title{Create the setup page}
\usage{
jvSetupPage(title, ...)
}
\arguments{
\item{title}{tile of the page}

\item{...}{elements to build the page from}
}
\description{
Create the setup page
}

