% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-layout.R
\name{jvTabbedContent}
\alias{jvTabbedContent}
\title{Create tabbed content}
\usage{
jvTabbedContent(..., selected = NULL, .list = NULL)
}
\arguments{
\item{...}{elements to create tabs from}

\item{selected}{boolean indicating which tab should be active}

\item{.list}{same as ..., but provided as a list}
}
\description{
Create tabbed content
}

