% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-layout.R
\name{jvTabNavItem}
\alias{jvTabNavItem}
\title{Create a tab nav item}
\usage{
jvTabNavItem(..., id, active = FALSE)
}
\arguments{
\item{...}{elements to pass}

\item{id}{html id}

\item{active}{boolean indicating if the tab item should be active.}
}
\description{
Create a tab nav item
}

