% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-layout.R
\name{jvTabPage}
\alias{jvTabPage}
\title{Create a tab page}
\usage{
jvTabPage(title, ..., id = makeHtmlId(title))
}
\arguments{
\item{title}{the title for the tab page}

\item{...}{elements to add as tabs}

\item{id}{html id of the tab page}
}
\description{
Create a tab page
}
