% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-layout.R
\name{jvTabs}
\alias{jvTabs}
\title{Create tabs}
\usage{
jvTabs(..., .list = NULL)
}
\arguments{
\item{...}{elements to pass.}

\item{.list}{same as ..., but passed as a list.}
}
\description{
Create tabs
}

