% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{listInstalledPackages}
\alias{listInstalledPackages}
\title{List of all installed packages used in Alteryx's Predictive tools}
\usage{
listInstalledPackages(svnDir = getOption("alteryx.svndir"), rVersion = NULL)
}
\arguments{
\item{svnDir}{Path to the local copy of a SVN branch of Alteryx.}

\item{rVersion}{The version of R to use as the basis of package installation.
For a completely new version of R, this will likely be the last version.}
}
\description{
A vector of all added packages in an SVN installation of a version of R.
This is done by reading the Readme.txt file in the /3rdParty/R/Installer
directory. The packages are in a named list identifying those from CRAN
and those from Alteryx
}
