% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{loveamongthechickens}
\alias{loveamongthechickens}
\title{"Love Among the Chickens"}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6543 rows and 3 columns.}
\source{
\url{http://www.gutenberg.org/cache/epub/3829/pg3829.txt}
#' @format A `tbl_df`, `tbl`, and `data.frame`
}
\usage{
loveamongthechickens
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"Love Among the Chickens", a book that launched his career.
}
\keyword{datasets}
