% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awesome-dashboard.R
\name{makeAwesomeDashboard}
\alias{makeAwesomeDashboard}
\title{Make awesome test dashboard}
\usage{
makeAwesomeDashboard(svnDir = getOption("alteryx.svndir"))
}
\description{
Make awesome test dashboard
}

