% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-list.R
\name{makeHtmlPageOfPackages}
\alias{makeHtmlPageOfPackages}
\title{Make Html Page for list of packages.}
\usage{
makeHtmlPageOfPackages(svnDir = getOption("alteryx.svndir"), saveTo = NULL)
}
\arguments{
\item{svnDir}{svn directory}

\item{saveTo}{file to save page to}
}
\description{
Make Html Page for list of packages.
}
