% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-layout.R
\name{makeJsVariable}
\alias{makeJsVariable}
\title{Make a javascript variable}
\usage{
makeJsVariable(x, name)
}
\arguments{
\item{x}{json object}

\item{name}{name to provide for the javascript variable}
}
\description{
Make a javascript variable
}
