% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-ui.R
\name{makeSubSection}
\alias{makeSubSection}
\title{Make a subsection}
\usage{
makeSubSection(subsection, config, ...)
}
\arguments{
\item{subsection}{subsection}

\item{config}{config}

\item{...}{additional arguments}
}
\description{
Make a subsection
}

