% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-workflow.R
\name{makeWorkflow}
\alias{makeWorkflow}
\title{Make a workflow}
\usage{
makeWorkflow(template, data, config, inputs_id, config_id, comment_id = NULL,
  comment = NULL, outFile, collapse = ",")
}
\arguments{
\item{template}{xml file to use as template for the workflow}

\item{data}{data to pass to the template}

\item{config}{configuration to use}

\item{inputs_id}{inputs_id}

\item{config_id}{config_id}

\item{comment_id}{comment_id}

\item{comment}{comment}

\item{outFile}{outFile}

\item{collapse}{collapse}
}
\description{
Make a workflow
}

