% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-workflow.R
\name{makeWorkflow2}
\alias{makeWorkflow2}
\title{Make an Alteryx workflow}
\usage{
makeWorkflow2(template, repl, outFile, collapse = ",")
}
\arguments{
\item{template}{xml file to use as template for the workflow}

\item{repl}{replacement to provide}

\item{outFile}{outFile}

\item{collapse}{collapse}
}
\description{
Make an Alteryx workflow
}

