% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{mike}
\alias{mike}
\title{Mike}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15160 rows and 3 columns.}
\source{
\url{http://www.gutenberg.org/cache/epub/7423/pg7423.txt}
#' @format A `tbl_df`, `tbl`, and `data.frame`
}
\usage{
mike
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"Mike", the work that introduced us to Mike (and a whole lot of cricketing jargon).
}
\keyword{datasets}
