% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{mikeandpsmith}
\alias{mikeandpsmith}
\title{Mike and Psmith}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/10586/pg10586.txt}
}
\usage{
mikeandpsmith
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"Mike and Psmith", a story  about two friends at a boarding school.
}
\keyword{datasets}
