% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/righthojeeves.R
\docType{data}
\name{mymanjeeves}
\alias{mymanjeeves}
\title{"My Man Jeeves"}
\format{A character vector with 6886 elements}
\source{
\url{https://www.gutenberg.org/files/8164/8164.txt}
}
\usage{
mymanjeeves
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's 1909
novel "My Man Jeeves".
}
\keyword{datasets}
