% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{notgeorgewashington}
\alias{notgeorgewashington}
\title{Not George Washington: A Smi-autobiographical Novel}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/7230/pg7230.txt}
}
\usage{
notgeorgewashington
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
semi-autobiographical novel.
}
\keyword{datasets}
