% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseBirds.R
\name{ParseBirds}
\alias{ParseBirds}
\title{ParseBirds}
\usage{
ParseBirds(x)
}
\arguments{
\item{birdstring}{Character string containing bird name. This should be either
the english common name or scientific name. If species cannot be ascertained,
the best-resolution taxa will be provided.}
}
\value{
A tibble with columns primary_common_name, order, phylum, genus,
species, and scientific_name.
}
\description{
Cleans and maps bird names to their English common name, scientific name and
taxonomy. Nomenclature is designated according to BirdLife.
}
