% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{piccadillyjim}
\alias{piccadillyjim}
\title{"Piccadilly Jim"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/2005/pg2005.txt}
}
\usage{
piccadillyjim
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
rollicking collection of four novels.
}
\keyword{datasets}
