% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{psmithinthecity}
\alias{psmithinthecity}
\title{Psmith in the City}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/6753/pg6753.txt}
}
\usage{
psmithinthecity
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
narrative revolving around Mike Jackson, Psmith, and cricketing dreams dashed by fate, or are they?
}
\keyword{datasets}
