% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{psmithjournalist}
\alias{psmithjournalist}
\title{Psmith, Journalist}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/2607/pg2607.txt}
}
\usage{
psmithjournalist
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
contiuing saga of Mike and Psmith and more cricket.
}
\keyword{datasets}
