% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readAndWriteMacro}
\alias{readAndWriteMacro}
\title{Read and write a macro}
\usage{
readAndWriteMacro(pluginDir = ".")
}
\arguments{
\item{pluginDir}{path to plugin directory}
}
\description{
Read and write a macro
}
