% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginGui.R
\name{renderPluginWidgets}
\alias{renderPluginWidgets}
\title{Render plugin widgets}
\usage{
renderPluginWidgets(pluginDir = ".", htmlFile = NULL, overrides = NULL,
  wrapInDiv = FALSE)
}
\arguments{
\item{pluginDir}{path to plugin directory}

\item{htmlFile}{layout file to use}

\item{overrides}{overrides if any}

\item{wrapInDiv}{boolean indicating if every widget should be wrapped in a div.}
}
\description{
Render plugin widgets
}
