% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{righthojeeves}
\alias{righthojeeves}
\title{Right Ho, Jeeves}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/10554/pg10554.txt}
}
\usage{
righthojeeves
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
classic collection of tales revolving around Bertie Wooster and the unflappable Jeeves.
}
\keyword{datasets}
