% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awesome-dashboard.R
\name{runAllSamplesAndTests}
\alias{runAllSamplesAndTests}
\title{Run all samples and QA tests.}
\usage{
runAllSamplesAndTests(svnDir = getOption("alteryx.svndir"))
}
\description{
Run all samples and QA tests.
}

