% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{runTests2}
\alias{runTests2}
\title{Run tests}
\usage{
runTests2(pluginDir = ".", build_doc = FALSE, testDir = "Tests")
}
\arguments{
\item{pluginDir}{path to the plugin directory}

\item{build_doc}{boolean indicating if docs should be built}

\item{testDir}{name of the tests directory to save results to}
}
\description{
Run tests
}
