% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWorkflow.R
\name{runWorkflow2}
\alias{runWorkflow2}
\title{Run an Alteryx workflow}
\usage{
runWorkflow2(file)
}
\arguments{
\item{file}{parameter to pass on to \code{AlteryxEngineCmd}}
}
\description{
Run an Alteryx workflow
}

