% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnfunctions.R
\name{saveReadme}
\alias{saveReadme}
\title{Save readme}
\usage{
saveReadme(save = TRUE, rVersion = getRversion())
}
\arguments{
\item{save}{whether or not to save the readme}

\item{rVersion}{version of R}
}
\description{
Save readme
}

