% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnfunctions.R
\name{svnInstallPackages}
\alias{svnInstallPackages}
\title{Install packages to SVN}
\usage{
svnInstallPackages(pkg, svnDir = getOption("alteryx.svndir"),
  rVersion = getRversion())
}
\arguments{
\item{pkg}{package to install}

\item{svnDir}{path to svn directory}

\item{rVersion}{version of R}
}
\description{
Install packages to SVN
}

