% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{talesofstaustins}
\alias{talesofstaustins}
\title{Tales of St. Austins}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/6980/pg6980.txt}
}
\usage{
talesofstaustins
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
collection of short stories and essays set in a public school, revolving around rugby
and cricket.
}
\keyword{datasets}
