% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{theadventuresofsally}
\alias{theadventuresofsally}
\title{"The Adventures of Sally"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/files/7464/7464-0.txt}
}
\usage{
theadventuresofsally
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
novel about the travails and tribulations of a heiress.
}
\keyword{datasets}
