% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{theclickingofcuthbert}
\alias{theclickingofcuthbert}
\title{"The Clicking of Cuthbert"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/7028/pg7028.txt}
}
\usage{
theclickingofcuthbert
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
classic fare that involves golf, and more golf.
}
\keyword{datasets}
