% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{thecomingofbill}
\alias{thecomingofbill}
\title{"The Coming of Bill"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/6880/pg6880.txt}
}
\usage{
thecomingofbill
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"serious" novel.
}
\keyword{datasets}
