% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{thegemcollector}
\alias{thegemcollector}
\title{"The Gem Collector"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/8931/pg8931.txt}
}
\usage{
thegemcollector
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
tale of Spike's misdeeds and eleventh-hour rescues that set the world right.
}
\keyword{datasets}
