% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{thegirlontheboat}
\alias{thegirlontheboat}
\title{"The Girl on the Boat"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/20717/pg20717.txt}
}
\usage{
thegirlontheboat
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
novek about red-haired, dog-loving "Billie".
}
\keyword{datasets}
