% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{thegoldbat}
\alias{thegoldbat}
\title{"The Gold Bat"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/20717/pg20717.txt}
}
\usage{
thegoldbat
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
novel set in Wrykyn School.
}
\keyword{datasets}
