% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{theheadofkays}
\alias{theheadofkays}
\title{The Head of Kays}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/6877/pg6877.txt}
}
\usage{
theheadofkays
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
novel set in a public school beset with the usual panoply of unruly pupils.
}
\keyword{datasets}
