% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{theintrusionofjimmy}
\alias{theintrusionofjimmy}
\title{The Intrusion of Jimmy}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/4075/pg4075.txt}
}
\usage{
theintrusionofjimmy
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
masterful treatment of romantic farce.
}
\keyword{datasets}
