% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{thelittlenugget}
\alias{thelittlenugget}
\title{"The Little Nugget"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/6683/pg6683.txt}
}
\usage{
thelittlenugget
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
1913 coming-of-age light comic novel set in a prep school.
}
\keyword{datasets}
