% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{thelittlewarrior}
\alias{thelittlewarrior}
\title{The Little Warrior}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/6837/pg6837.txt}
}
\usage{
thelittlewarrior
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
romantic tale set on both sides of the Atlantic with uncommon twists and turns.
}
\keyword{datasets}
