% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{themanupstairs}
\alias{themanupstairs}
\title{"The Man Upstairs and Other Stories"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/6768/pg6768.txt}
}
\usage{
themanupstairs
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"The Man Upstairs"
}
\keyword{datasets}
