% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{themanwithtwoleftfeet}
\alias{themanwithtwoleftfeet}
\title{"The Man with Two Left Feet"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/7471/pg7471.txt}
}
\usage{
themanwithtwoleftfeet
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
collection of short stories that might introduce Jeeves.
}
\keyword{datasets}
