% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{theprinceandbetty}
\alias{theprinceandbetty}
\title{"The Prince and Betty"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/6955/pg6955.txt}
}
\usage{
theprinceandbetty
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
novel about a scheming billionaire, his daughter's love interest, and casino in the works.
}
\keyword{datasets}
