% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{theswooporhowclarencesavedengland}
\alias{theswooporhowclarencesavedengland}
\title{"The Swoop or how Clarence Saved England"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/7050/pg7050.txt}
}
\usage{
theswooporhowclarencesavedengland
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"The Swoop!, or How Clarence Saved England".
}
\keyword{datasets}
