% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{threemenandamaid}
\alias{threemenandamaid}
\title{"Three Men and a Maid"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/6836/pg6836.txt}
}
\usage{
threemenandamaid
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
"A Girl on a Boat", initially published in the United States as "Three Men and a Maid".
I have kept it in the database because the length seems to be different from that of
"A Girl on a Boat" and I'd be curious to see why.
}
\keyword{datasets}
