% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{uneasymoney}
\alias{uneasymoney}
\title{Uneasy Money}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/6684/pg6684.txt}
}
\usage{
uneasymoney
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
novel about a man finding it hard to give money away to those hard-done by a will,
and is set in Long Island (NY).
}
\keyword{datasets}
