% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFromGithub.R
\name{updatePackageFromGithub}
\alias{updatePackageFromGithub}
\alias{updatePluginFromGithub}
\title{Update SVN with R package or HtmlPlugin from Github}
\usage{
updatePackageFromGithub(name, svnDir = getOption("alteryx.svndir"))

updatePluginFromGithub(name, svnDir = getOption("alteryx.svndir"))
}
\arguments{
\item{name}{name of the package or plugin to update.}

\item{svnDir}{path to the root of the svn directory.}
}
\description{
Update SVN with R package or HtmlPlugin from Github
}
\examples{
# Set path to your SVN directory using options(alteryx.svndir = <path>)
updatePackageFromGithub('AlteryxPredictive')
updatePluginFromSvn('Linear_Regression')
}

