% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{williamtelltoldagain}
\alias{williamtelltoldagain}
\title{William Tell Told Again}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{http://www.gutenberg.org/cache/epub/7298/pg7298.txt}
}
\usage{
williamtelltoldagain
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
retelling of the legendary tales of William Tell, in prose and verse.
}
\keyword{datasets}
