% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginGui.R
\name{writeGuiHtml}
\alias{writeGuiHtml}
\title{Write Gui HTML for plugin}
\usage{
writeGuiHtml(pluginDir = ".", htmlFile = NULL, overrides = NULL)
}
\arguments{
\item{pluginDir}{directory containing the plugin.}

\item{htmlFile}{html file to write the Gui to}

\item{overrides}{a list of overrides for widgets}
}
\description{
Write Gui HTML for plugin
}
