% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-ui.R
\name{writeLayoutFromYamlSpec}
\alias{writeLayoutFromYamlSpec}
\title{Write layout from a yaml spec file}
\usage{
writeLayoutFromYamlSpec(pluginDir = ".", displayRules = c(), ...)
}
\arguments{
\item{pluginDir}{path to root folder of the plugin.}

\item{displayRules}{list of display rules to specify.}

\item{dataItems}{arguments to pass to jvMakeDataItemsToInitialize.}
}
\description{
Write layout from a yaml spec file
}

