% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-ui.R
\name{writeLayoutFromYamlSpec2}
\alias{writeLayoutFromYamlSpec2}
\title{Write layout file from a YAML spec}
\usage{
writeLayoutFromYamlSpec2(pluginDir = ".")
}
\arguments{
\item{pluginDir}{path to plugin}
}
\description{
Write layout file from a YAML spec
}

