% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginConfig.R
\name{writePluginConfig}
\alias{writePluginConfig}
\alias{yxmc2PluginConfig}
\title{Make plugin configuration xml file from macro}
\usage{
writePluginConfig(yxmc, saveToFile = NULL)

yxmc2PluginConfig(yxmc, saveToFile = NULL)
}
\arguments{
\item{yxmc}{path to macro}

\item{saveToFile}{file to save config.xml to}
}
\description{
Make plugin configuration xml file from macro
}
