% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{writeRPluginIni}
\alias{writeRPluginIni}
\title{Write RPluginIni}
\usage{
writeRPluginIni(revo = FALSE, replace = FALSE)
}
\arguments{
\item{revo}{boolean indicating if xdf inputs are to be supported.}

\item{replace}{boolean indicating if the existing file should be overwritten.}
}
\description{
The \code{RPluginSettings.ini} file enables Alteryx to find the R engine to
work with. This function allows one to customize the ini file and support
working with R installs that are not provided by the predictive installer.
}
