% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{adamselindistress}
\alias{adamselindistress}
\title{"A Damsel in Distress"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{https://www.gutenberg.org/files/2233/2233-8.txt}
}
\usage{
adamselindistress
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's
novel "A Damsel in Distress". First published in 1919, the plot revolves
around golf-loving #' American composer George Bevan who falls in love with
a mysterious young lady who takes refuge in his taxicab one day. When he
later tracks her down to a romantic rural manor, mistaken identity leads to
all manner of brouhaha.
}
\keyword{datasets}
