% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{copyXDFFiles}
\alias{copyXDFFiles}
\title{Copy XDF files from SVN}
\usage{
copyXDFFiles(svnDir = getOption("alteryx.svndir"), rVersion = getRversion())
}
\arguments{
\item{svnDir}{svn directory to copy from.}

\item{rVersion}{string indicating version of R.}
}
\description{
Copy XDF files from SVN
}
