% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDashboard.R
\name{createTestDashboard}
\alias{createTestDashboard}
\title{Create a dashboard of test results}
\usage{
createTestDashboard(plugins)
}
\arguments{
\item{plugins}{plugins to get test results for}
}
\description{
Create a dashboard of test results
}

