% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractConfiguration.R
\name{extractConfiguration}
\alias{extractConfiguration}
\alias{yxmc2yaml}
\title{Extract question constants and their configuration from macro}
\usage{
extractConfiguration(yxmc, asYaml = F)

yxmc2yaml(yxmc, asYaml = F)
}
\arguments{
\item{yxmc}{path to yxmc file to extract configuration from.}

\item{asYaml}{boolean indicating whether output should be converted to yaml.}
}
\description{
Extract question constants and their configuration from macro
}
\examples{
yxmc <- system.file("templates", "sample1.yxmc", package = 'jeeves')
d <- extractConfiguration(yxmc)
}

