% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupSequences.R
\name{GroupSequences}
\alias{GroupSequences}
\title{GroupSequences}
\usage{
GroupSequences(aln, snp_threshold = 0)
}
\arguments{
\item{aln}{Nucleotide sequence alignment of class DNAbin or phyDat}

\item{snp_threshold}{Integer (n>=0) indicating the maximum within-group
Hamming distance.}
}
\value{
A tibble with columns sequence_name and sequence_group
}
\description{
Calculate Hamming distance of a nucleotide sequence alignment and allocate
groups according to a threshold.
}
