% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{installAllPackages}
\alias{installAllPackages}
\title{Install All Needed Packages}
\usage{
installAllPackages(dev = TRUE, rVersion = NULL)
}
\arguments{
\item{dev}{Boolean indicating if dev versions of packages should be installed.}

\item{rVersion}{The version of R to use as the basis of package installation.
This optional argument will typically be used when moving to a new version of
R, when the natural source of packages were those packages in the current
version of R used by Alteryx.}
}
\description{
Install All Needed Packages
}
