% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeeves-package.R
\docType{package}
\name{jeeves-package}
\alias{jeeves}
\alias{jeeves-package}
\title{jeeves: Tools for reproducible research workflows in R}
\description{
An opinionated collection of tools for creating, managing and sharing reproducible research projects in R. Create standard project structures, manage dependencies, and generate reports with ease.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/venpopov/jeeves}
  \item \url{https://venpopov.github.io/jeeves/}
  \item Report bugs at \url{https://github.com/venpopov/jeeves/issues}
}

}
\author{
\strong{Maintainer}: Ven Popov \email{vencislav.popov@gmail.com} (\href{https://orcid.org/0000-0002-8073-4199}{ORCID})

}
\keyword{internal}
