% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeevesdata.R
\docType{data}
\name{mymanjeeves}
\alias{mymanjeeves}
\title{"My Man Jeeves"}
\format{A `tbl_df`, `tbl`, and `data.frame`}
\source{
\url{https://www.gutenberg.org/files/8164/8164.txt}
}
\usage{
mymanjeeves
}
\description{
A data set containing the complete text of Pelham Grenville Wodehouse's 1919
"My Man Jeeves", a collection of short stories that feature Jeeves,
Bertie Wooster, and Reggie Pepper (an early prototype for Bertie Wooster).
}
\keyword{datasets}
