% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseDate.R
\name{ParseDate}
\alias{ParseDate}
\title{ParseDate}
\usage{
ParseDate(date_str)
}
\arguments{
\item{date_str}{date string. May take yyyy-mm-dd, dd-mm-yyyy, yyyy-mm, mm-yyyy
or yyyy formats. Caution advised if using mm-dd-yyyy.}
}
\value{
A list-column of formats including ymd, ym, y, decimal and precision.
}
\description{
Format a date string with variable precision/delimiters. Output a
}
\examples{
df <- tibble(collection_date = c("2025-08-14",
 "14-08-2025", "2025-08", "08-2025", "2025"))

df_nested <- df \%>\%
 mutate(parsed_info = ParseDate(collection_date)) \%>\%
 unnest(parsed_info)

}
