% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{updatePackageFromGithub}
\alias{updatePackageFromGithub}
\title{Update R package from Github}
\usage{
updatePackageFromGithub(pkg, svnDir = getOption("alteryx.svndir"))
}
\arguments{
\item{pkg}{name of the package.}

\item{svnDir}{path to the root of the svn directory.}
}
\description{
Update R package from Github
}
\examples{
updatePackageFromGithub('AlteryxPredictive')
}

