% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFromGithub.R
\name{updatePackageFromGithub}
\alias{updatePackageFromGithub}
\alias{updatePluginFromGithub}
\title{Update SVN with R package or HtmlPlugin from Github}
\usage{
updatePackageFromGithub(name, install = FALSE,
  svnDir = getOption("alteryx.svndir"), ayxDir = getOption("alteryx.path"))

updatePluginFromGithub(name, svnDir = getOption("alteryx.svndir"),
  install = FALSE, ayxDir = getOption("alteryx.path"))
}
\arguments{
\item{name}{name of the package or plugin to update.}

\item{install}{boolean indicating if it should install to the alteryx dir.}

\item{svnDir}{path to the root of the svn branch directory.}

\item{ayxDir}{path to alteryx directory.}
}
\description{
Update SVN with R package or HtmlPlugin from Github
}
\section{Update Plugin from Github}{

This function tries to automate the following manual sequence of steps.
\enumerate{
  \item Download zip file from github.
  \item Unzip it and rename it (remove the -master suffix).
  \item Delete the Extras subfolder contained in it.
  \item Copy the folder over to SVN.
}
}

\examples{
# Set path to your SVN directory using options(alteryx.svndir = <path>)
\dontrun{
  updatePackageFromGithub('AlteryxPredictive')
  updatePluginFromSvn('Linear_Regression')
}
}
